Removed quote and code block command, Moved link command into link package and called from rich-text package
